separated cisco_handler_base to separate classes.
