+++ cloudshell/tests/networking/cisco/cisco_handler_base/connectivity_methods/test_apply_connectivity_changes.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 from mock import MagicMock
 import re
-from cloudshell.networking.cisco.cisco_handler_base import CiscoHandlerBase
+from cloudshell.networking.cisco.cisco_send_command_operations import CiscoHandlerBase
 
 
 class TestCiscoHandlerBase(TestCase):
