+++ cloudshell/tests/networking/cisco/cisco_handler_base/save_restore_methods/test_restore_configuration_fixed_bugs.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 from mock import MagicMock
 import re
-from cloudshell.networking.cisco.cisco_handler_base import CiscoHandlerBase
+from cloudshell.networking.cisco.cisco_send_command_operations import CiscoHandlerBase
 
 __author__ = 'CoYe'
